requirements.txt
Flask==2.2.2
redis==4.5.1
python-dotenv==1.0.0